refine switching

+++ scripts/templates/links.sh
@@ -2,6 +2,10 @@
 cd $(dirname $0)
 world=${1:-RestWorld}
 target="$HOME/clarity/home/saves/$world/datapacks/restworld/data/v3/functions"
+if [ test ! -d "$target"]; then
+    echo No such directory: $target 1>&2
+    exit 1
+fi
 rm */.f
 for f in *; do
     if [ -d $f -a -d $target/$f ]; then
